Retoques en html y eliminación de clases css

+++ resources/views/comunidades/index.blade.php
@@ -4,12 +4,10 @@
 @endsection
 @section('contenido')
     <div class="spinner"></div>
-    <div class="hidden table-size-optima altoMaximo">
+    <div class="hidden table-size-optima">
         @if (Auth::check())
-            <div class="row ">
-                @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'comunidades.parciales.buscar'])
-                @include('comun.plantillaOperacionesIndex',['tabla'=>'comunidades','accion'=>'Nueva'])
-            </div>
+            @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'comunidades.parciales.buscar'])
+            @include('comun.plantillaOperacionesIndex',['tabla'=>'comunidades','accion'=>'Nueva'])
             @if(!$comunidades->isEmpty())
                 @foreach ($comunidades as $comunidad)
                     <div>
@@ -33,33 +31,35 @@
                                         </i>
                                     </a>
                                     @if ((Auth::user()->roles->peso)>=config('opciones.roles.administrador')){{--Administrador --}}
-                                    {!! FORM::open(array('route' => array('comunidades.destroy',
-                                    $comunidad->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar')))!!}
-                                    <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
-                                            @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                            class="pull-right lanzarModal simpleModal"
-                                            data-modal_sin_etiqueta="true"
-                                            data-modal_ancho="330"
-                                            data-modal_cabecera_color_fondo='rgba(255,0,0,.9)'
-                                            data-modal_cabecera_color_texto='#ffffff'
-                                            data-modal_cuerpo_color_fondo='rgba(255,255,255,.9)'
-                                            data-modal_cuerpo_color_texto='"#ffffff'
-                                            data-modal_pie_color_fondo='#400090'
-                                            data-modal_pie_color_texto='"#ffffff'
-                                            data-modal_posicion_vertical="220"
-                                            data-titulo="BORRAR"
-                                            data-pie="true"
-                                            data-descripcion="¿Seguro que deseas eliminar esta comunidad?<br><h3><strong>{{$comunidad->comunidad}}</strong></h3>"
-                                            data-footer="true"
-                                            @endif >
-                                        <i class='glyphicon glyphicon-trash full-Width'>
-                                            <div>Borrar</div>
-                                        </i>
-                                    </button>
-                                    @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                        @include ("comun.plantillaBorrado")
+                                    @if($comunidad->activa)
+                                        {!! FORM::open(array('route' => array('comunidades.destroy',
+                                        $comunidad->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar')))!!}
+                                        <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
+                                                @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                                class="pull-right lanzarModal simpleModal"
+                                                data-modal_sin_etiqueta="true"
+                                                data-modal_ancho="330"
+                                                data-modal_cabecera_color_fondo='rgba(255,0,0,.9)'
+                                                data-modal_cabecera_color_texto='#ffffff'
+                                                data-modal_cuerpo_color_fondo='rgba(255,255,255,.9)'
+                                                data-modal_cuerpo_color_texto='"#ffffff'
+                                                data-modal_pie_color_fondo='#400090'
+                                                data-modal_pie_color_texto='"#ffffff'
+                                                data-modal_posicion_vertical="220"
+                                                data-titulo="BORRAR"
+                                                data-pie="true"
+                                                data-descripcion="¿Seguro que deseas eliminar esta comunidad?<br><h3><strong>{{$comunidad->comunidad}}</strong></h3>"
+                                                data-footer="true"
+                                                @endif >
+                                            <i class='glyphicon glyphicon-trash full-Width'>
+                                                <div>Borrar</div>
+                                            </i>
+                                        </button>
+                                        @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                            @include ("comun.plantillaBorrado")
+                                        @endif
+                                        {!! FORM::close() !!}
                                     @endif
-                                    {!! FORM::close() !!}
                                     @endif
                                 </th>
                             </tr>
@@ -135,6 +135,7 @@
                         </table>
                     </div>
                 @endforeach
+                {!! $comunidades->appends(Request::only(['comunidad','pais']))->render()!!}
             @else
                 <div class="">
                     <div class="alert alert-info" role="alert">
@@ -142,12 +143,6 @@
                     </div>
                 </div>
             @endif
-            <div class="row paginationBlock">
-                {!! $comunidades->appends(Request::only(['comunidad','pais']))->render()
-                !!}{{-- Poner el paginador --}}
-                {{--{!! $comunidades->appends(Request::only(['comunidad','esPropia','secretariado','pais']))->render()
-                !!}--}}{{-- Poner el paginador --}}
-            </div>
         @else
             @include('invitado')
         @endif
