+++ resources/views/localidades/index.blade.php
@@ -4,30 +4,30 @@
 @endsection
 @section('contenido')
     <div class="spinner"></div>
-    <div class="hidden table-size-optima altoMaximo">
+    <div class="hidden table-size-optima">
         @if (Auth::check())
-            <div class="row ">
-                @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'localidades.parciales.buscar'])
-                @include('comun.plantillaOperacionesIndex',['tabla'=>'localidades','accion'=>'Nueva'])
-                @if(!$localidades->isEmpty())
-                    @foreach ($localidades as $localidad)
-                        <div class="full-Width">
-                            <table class="table-viaoptima table-striped">
-                                <caption
-                                        class="label-default text-center  @if(!$localidad->activo) foreground-disabled @else alert-warning @endif">
-                                    {{ $localidad->localidad }}
-                                </caption>
-                                <thead>
-                                <tr @if(!$localidad->activo) class="background-disabled" @endif>
-                                    <th colspan="2" class="text-right">
-                                        <a title="Editar"
-                                           href="{{route('localidades.edit',array('id'=>$localidad->id))}}">
-                                            <i class="glyphicon glyphicon-edit">
-                                                <div>Editar</div>
-                                            </i>
-                                        </a>
 
-                                        @if ((Auth::user()->roles->peso)>=config('opciones.roles.administrador')){{--Administrador --}}
+            @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'localidades.parciales.buscar'])
+            @include('comun.plantillaOperacionesIndex',['tabla'=>'localidades','accion'=>'Nueva'])
+            @if(!$localidades->isEmpty())
+                @foreach ($localidades as $localidad)
+                    <div class="full-Width">
+                        <table class="table-viaoptima table-striped">
+                            <caption
+                                    class="label-default text-center  @if(!$localidad->activo) foreground-disabled @else alert-warning @endif">
+                                {{ $localidad->localidad }}
+                            </caption>
+                            <thead>
+                            <tr @if(!$localidad->activo) class="background-disabled" @endif>
+                                <th colspan="2" class="text-right">
+                                    <a title="Editar"
+                                       href="{{route('localidades.edit',array('id'=>$localidad->id))}}">
+                                        <i class="glyphicon glyphicon-edit">
+                                            <div>Editar</div>
+                                        </i>
+                                    </a>
+                                    @if ((Auth::user()->roles->peso)>=config('opciones.roles.administrador')){{--Administrador --}}
+                                    @if($localidad->activo)
                                         {!! FORM::open(array('route' => array('localidades.destroy',
                                         $localidad->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar')))!!}
                                         <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
@@ -55,39 +55,36 @@
                                             @include ("comun.plantillaBorrado")
                                         @endif
                                         {!! FORM::close() !!}
-                                        @endif
-                                    </th>
-                                </tr>
-                                </thead>
-                                <tbody>
-                                <tr @if(!$localidad->activo) class="foreground-disabled" @endif>
-                                    <td class="table-autenticado-columna-1">País:</td>
-                                    <td>{{ $localidad->pais }}</td>
-                                </tr>
-                                <tr @if(!$localidad->activo) class="foreground-disabled" @endif>
-                                    <td>Provincia:</td>
-                                    <td>{{ $localidad->provincia }}</td>
-                                </tr>
-                                </tbody>
-                            </table>
-                        </div>
-                    @endforeach
-                @else
-                    <div class="clearfix">
-                        <div class="alert alert-info" role="alert">
-                            <p><strong>¡Aviso!</strong> No se ha encontrado ninguna localidad que
-                                listar.</p>
-                        </div>
+                                    @endif
+                                    @endif
+                                </th>
+                            </tr>
+                            </thead>
+                            <tbody>
+                            <tr @if(!$localidad->activo) class="foreground-disabled" @endif>
+                                <td class="table-autenticado-columna-1">País:</td>
+                                <td>{{ $localidad->pais }}</td>
+                            </tr>
+                            <tr @if(!$localidad->activo) class="foreground-disabled" @endif>
+                                <td>Provincia:</td>
+                                <td>{{ $localidad->provincia }}</td>
+                            </tr>
+                            </tbody>
+                        </table>
                     </div>
-                @endif
-                <div class="row paginationBlock">
-                    {!! $localidades->appends(Request::only(['localidad']))->render()
-                    !!}{{-- Poner el paginador --}}
+                @endforeach
+                {!! $localidades->appends(Request::only(['localidad']))->render()!!}
+            @else
+                <div class="clearfix">
+                    <div class="alert alert-info" role="alert">
+                        <p><strong>¡Aviso!</strong> No se ha encontrado ninguna localidad que
+                            listar.</p>
+                    </div>
                 </div>
-                @else
-                    @include('comun.guestGoHome')
-                @endif
-            </div>
+            @endif
+        @else
+            @include('comun.guestGoHome')
+        @endif
     </div>
 @endsection
 @section('js')
